+++ TheMET/ru.lproj/Localizable.strings
@@ -20,6 +20,7 @@
 "catalog_screen_departments_section_title" = "Департамент";
 "catalog_screen_cultures_section_title" = "Культура";
 "catalog_screen_feaured_artists_section_title" = "Избранные художники";
+"about_met_screen_title" = "O MET";
 "catalog.loading_failed" = "Каталог не может быть загружен  \nУбедитесь, что вы подключены к Wi-Fi или сети мобильного интернета";
 "favories_screen_title" = "Избранное";
 "about_app_screen_title" = "О приложении";
